% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{MR Barplot}
\usage{
barplot(obj, ...)
}
\arguments{
\item{obj}{soem object}

\item{...}{more params}
}
\description{
Barplot for MR object
}
\details{
Method ..
}
\author{
Junjie Zheng
}

