% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplotMR}
\alias{barplotMR}
\alias{barplotMR.b2}
\alias{barplotMR.between}
\alias{barplotMR.bymrocalc}
\alias{barplotMR.mrocalc}
\title{MR Barplot}
\usage{
barplotMR(obj, ...)

\method{barplotMR}{mrocalc}(obj, ...)

\method{barplotMR}{bymrocalc}(obj, g1.level = NULL, g2.level = "_MULTI",
  ...)

\method{barplotMR}{between}(obj, ...)

\method{barplotMR}{b2}(obj, g1.level = NULL, ...)
}
\arguments{
\item{obj}{soem object}

\item{...}{more params}

\item{g1.level}{vector of subset variable 1 levels to show}

\item{g2.level}{vector of subset variable 2 levels to show}
}
\description{
Barplot for MR object
}
\details{
Method ..
}
\section{Methods (by class)}{
\itemize{
\item \code{mrocalc}: method for class `mrocalc`

\item \code{bymrocalc}: method for class `bymrocalc`

\item \code{between}: method for class `between`

\item \code{b2}: method for class `b2`
}}
\author{
Junjie Zheng
}

