% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bymro.R
\name{byMRO}
\alias{byMRO}
\title{Construct MRO subset by variable}
\usage{
byMRO(mro.obj, formula, FUN, ...)
}
\arguments{
\item{mro.obj}{an mro object}

\item{formula}{formula for subsetting}

\item{FUN}{the function to apply to subsets}

\item{...}{additional arguments}
}
\value{
an object with classes of \code{by} and \code{bymrocalc}
}
\description{
Constructs a multiple response object (MRO) subset by another explanatory variable.
}
\author{
Junjie Zeng
}
