% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfactor.R
\name{mfactor}
\alias{mfactor}
\title{Create a multi-response factor variable}
\usage{
mfactor(x, levels = names(x), indicator = NULL, na.value = 0L)
}
\arguments{
\item{x}{a dataframe of columns for each response option}

\item{levels}{vector of level names (defaults to names of \code{x})}

\item{indicator}{if values in \code{x} are not coercible to logical,
then indicator can either be the value used to indicate
a selection (e.g., 'yes'), or a function if a more complex
criteria is required (e.g., \code{function(x) x > 0}).}

\item{na.value}{usually, missing values in some columns will be treated
as \code{FALSE}/zero, but this can be overridden by specifying
an alternative value (1, or any other value). Note that
currently missing values are considered 'non-response',
and so are treated in exactly the same way as a 'zero' response.
If all values for a row are missing, then the result will always be \code{NA}.}
}
\value{
A vector of class \code{mfactor} where each item is a character
string of zeros and ones indicating, in order,
the response value associated with that row.
}
\description{
Create a multi-response factor variable
}
\examples{

}
\author{
Tom Elliott
}
