% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mroPara.R
\name{mroPara}
\alias{mroPara}
\title{Compute Para MRO for an object}
\usage{
mroPara(obj, conf.levels = 1.96, nonparallel = NULL)
}
\arguments{
\item{obj}{An object to compute things for}

\item{conf.levels}{Confidence level to use}

\item{nonparallel}{Should these things be parallel?}
}
\value{
An object of class \code{mrocalc}
}
\description{
Computes the MRO information for a para object ...
}
\author{
Junjie Zeng
}

