% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{seBinprops}
\alias{seBinprops}
\title{Independent Binomial Proportions}
\usage{
seBinprops(ns, phats)
}
\arguments{
\item{ns}{the number of observations in the independent groups}

\item{phats}{the proportions of TRUE/1's etc.}
}
\value{
an \code{ses.moecalc} object
}
\description{
Compute SEs for Independent Binomial Proportions
}
\author{
Junjie Zeng
}
