% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{seIndepSes}
\alias{seIndepSes}
\title{Independent Standard errors given}
\usage{
seIndepSes(ses)
}
\arguments{
\item{ses}{the standard errors}
}
\value{
an \code{ses.moecalc} object
}
\description{
Returns ses.moecalc for given SEs
}
\author{
Junjie Zeng
}
