% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{seMNprops}
\alias{seMNprops}
\title{Compute SE for Multinomial proportions}
\usage{
seMNprops(n, phat)
}
\arguments{
\item{n}{the number of observations in each group}

\item{phat}{the estimates proportions for each group}
}
\value{
an \code{ses.moecalc} object
}
\description{
SEs for Multinomial Proportions
}
\examples{
phat <- table(iris$Species) / nrow(iris)
seMNprops(nrow(iris), phat)
}
\author{
Junjie Zeng
}
