% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{seMRprops}
\alias{seMRprops}
\title{Multiple binary response}
\usage{
seMRprops(obj)
}
\arguments{
\item{obj}{something that can be turned into a matrix}
}
\value{
an \code{ses.moecalc} object
}
\description{
SE's for multiple binary response
}
\examples{
x <- data.frame(
 v1 = rbinom(20, 1, 0.8),
 v2 = rbinom(20, 1, 0.3),
 v3 = rbinom(20, 1, 0.5)
)
seMRprops(x)
}
\author{
Junjie Zeng
}
