% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{substrsplit}
\alias{substrsplit}
\title{Extract Common Name from variables}
\usage{
substrsplit(obj)
}
\arguments{
\item{obj}{It can be a vector or data frame, however, \code{substrsplit} is usually
used in the \code{iNZightMR} function.}
}
\value{
A list with common character and unique variable name respectively
}
\description{
Help mro variables extract common name out
}
\examples{
substrsplit(c("varx", "vary", "varz"))
}
