% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switcher.R
\name{switcher}
\alias{switcher}
\alias{switcher.b2}
\alias{switcher.between}
\alias{switcher.bymrocalc}
\alias{switcher.mrocalc}
\title{Switcher generic}
\usage{
switcher(obj, ...)

\method{switcher}{mrocalc}(obj, ...)

\method{switcher}{bymrocalc}(obj, ...)

\method{switcher}{between}(obj, ...)

\method{switcher}{b2}(obj, ...)
}
\arguments{
\item{obj}{mro familiy}

\item{\code{...}}{remain future use}
}
\description{
Switcher generic
}
\section{Methods (by class)}{
\itemize{
\item \code{mrocalc}: For mrocalc case

\item \code{bymrocalc}: For bymrocalc case

\item \code{between}: For between case

\item \code{b2}: For between and b2 case
}}

