% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{between}
\alias{between}
\title{Compute Between se's}
\usage{
between(bymro)
}
\arguments{
\item{bymro}{a bymro object}
}
\value{
something about between.
}
\description{
Between SEs
}
\author{
Junjie Zheng
}

