% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bymro.R
\name{byMRO}
\alias{byMRO}
\title{By MRO Constructor}
\usage{
byMRO(mro.obj, formula, FUN, ...)
}
\arguments{
\item{mro.obj}{an mro object}

\item{formula}{formula for subsetting}

\item{FUN}{the function to apply to subsets}

\item{...}{additional arguments}
}
\value{
a dataframe
}
\description{
Construct MRO by a given subset variable.
}
\author{
Junjie Zeng
}

