% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmro.R
\name{calcmissing}
\alias{calcmissing}
\alias{calcmissing.data.frame}
\alias{calcmissing.mro}
\title{Calculate missing observation summary}
\usage{
calcmissing(obj, ...)

\method{calcmissing}{data.frame}(obj, MRO.case = FALSE, print = TRUE, final = TRUE, ...)

\method{calcmissing}{mro}(obj, ...)
}
\arguments{
\item{obj}{An object}

\item{...}{additional arguments}

\item{MRO.case}{does something with rownames}

\item{print}{logical, should we print the thing?}

\item{final}{logical, is this final?}
}
\value{
Missing value object
}
\description{
Calculates the summary of missingness in a data set.
}
\section{Methods (by class)}{
\itemize{
\item \code{calcmissing(data.frame)}: Method for a dataframe

\item \code{calcmissing(mro)}: accepts a whole mr.object , which is first mro.mat, second element labels,
third element the input data frame.

}}
\examples{
calcmissing(census.at.school.5000[,1:20])
}
\seealso{
plotcombn
}
\author{
Junjie Zeng
}
