% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMR-package.R
\docType{package}
\name{iNZightMR-package}
\alias{iNZightMR-package}
\title{iNZightMR: Multiple Response Data Analysis}
\description{
The iNZightMR package provides a suite of functions which can be used
in the analysis of multiple response data. It is used in the
iNZight software package.
}
\details{
\tabular{ll}{
  Package: \tab iNZightMR\cr
  Type: \tab Package\cr
  Version: \tab 2.0\cr
  Date: \tab 2014-09-19\cr
  License: \tab GPL-2\cr
}
}
\author{
Junjie Zeng, Tom Elliott
}
\keyword{iNZight}
\keyword{multiple}
\keyword{response}
