% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMR-package.R
\docType{package}
\name{iNZightMR-package}
\alias{iNZightMR-package}
\title{iNZightMR: Multiple Response Data Analysis}
\description{
The iNZightMR package provides a suite of functions which can be used
in the analysis of multiple response data. It is used in the
iNZight software package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://inzight.nz}
  \item \url{https://inzightvit.github.io/iNZightMR/}
  \item Report bugs at \url{https://github.com/iNZightVIT/iNZightMR/issues}
}

}
\author{
Junjie Zeng, Tom Elliott
}
\keyword{iNZight}
\keyword{multiple}
\keyword{response}
