\name{iNZightMR}
\docType{package}
\title{
  Multiple Response Data Analysis
}
\description{
  The iNZightMR package provides a suite of functions which can be used
  in the analysis of multiple response data. It is used in the
  iNZight software package.
}
\details{
\tabular{ll}{
  Package: \tab iNZightMR\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-02-25\cr
  License: \tab GPL-2\cr
}
}
\author{
  Junjie Zeng, originally developed by Simon Potter.
  
  Maintainer: Junjie Zeng <jzen188@aucklanduni.ac.nz>
}
\keyword{ iNZight }
\seealso{
  \code{\link{iNZight}}, \code{\link{iNZightRegression}}
}
\examples{

}
