% Generated by roxygen2 (4.0.2): do not edit by hand
\name{iNZightMR}
\alias{iNZightMR}
\title{Create iNZightMR multiple response object (MRO)}
\usage{
iNZightMR(frm, data, Labels = NULL, inverse = substrsplit, ...)
}
\arguments{
\item{frm}{A formula}

\item{data}{A data}

\item{Labels}{default NULL}

\item{inverse}{A logical value}

\item{\code{...}}{Extra arguments like in \code{model.frame}}
}
\value{
The MRO list containing a multiple response binary matrix and input
  data source
}
\description{
\code{iNZightMR} returns the multiple response object (MRO) for stable
analysis and encapsulation purpose
}
\examples{
mr <- iNZightMR(online~onlinegame+onlinevideo+onlinemusic, data = CaS)
}

