% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iNZightMR.R
\name{iNZightMR}
\alias{iNZightMR}
\title{Create iNZightMR multiple response object (MRO)}
\usage{
iNZightMR(frm, data, Labels = NULL, inverse = FALSE, ...)
}
\arguments{
\item{frm}{A formula}

\item{data}{A data}

\item{Labels}{default NULL}

\item{inverse}{A logical value}

\item{...}{Extra arguments like in \code{model.frame}}
}
\value{
An \code{mro} object containing a multiple response binary matrix and input
  data source
}
\description{
Creates a multiple response object (MRO) containing binary reponse matrix as
well as the input data source. This can then be passed to other
multiple response functions, such as \code{barplotMR} or \code{mroPara}.
}
\examples{
\dontrun{
  mr <- iNZightMR(online~onlinegame+onlinevideo+onlinemusic, data = census.at.school.5000)
}
}
\seealso{
\link{barplotMR}, \link{mroPara}
}
\author{
Junjie Zeng
}
