% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmro.R
\name{plotcombn}
\alias{plotcombn}
\title{Missing Value plot}
\usage{
plotcombn(obj)
}
\arguments{
\item{obj}{a calcmissing object}
}
\value{
summarised info for plot
}
\description{
Plot of Missing Value combinations
}
\author{
Junjie Zeng
}

