% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{seCovs}
\alias{seCovs}
\title{Compute standard error for covariance matrix}
\usage{
seCovs(covs, addbase = FALSE)
}
\arguments{
\item{covs}{covariance matrix}

\item{addbase}{logical, is there a baseline?}
}
\value{
an ses.moecalc object
}
\description{
Compute the standard error information for a given covariance matrix.
}
\examples{
seCovs(cov(iris[,-5]))
}
\author{
Junjie Zeng
}
