% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.moecalc.R
\name{seMRprops}
\alias{seMRprops}
\title{Multiple binary response}
\usage{
seMRprops(obj)
}
\arguments{
\item{obj}{something that can be turned into a matrix}
}
\value{
an \code{ses.moecalc} object
}
\description{
SE's for multiple binary response
}
\author{
Junjie Zeng
}
