% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{substrsplit}
\alias{substrsplit}
\title{Extract Common Name from variables}
\usage{
substrsplit(obj)
}
\arguments{
\item{obj}{It can be a vector or data frame, however, \code{substrsplit} is usually
used in the \code{iNZightMR} function.}
}
\value{
A list with common character and unique variable name respectively
}
\description{
Help mro variables extrac common name out
}
\examples{
\dontrun{
mr <- iNZightMR(online~onlinegame+onlinevideo+onlinemusic,
                data = census.at.school.5000)
mroPara(mr)
mr2 <- iNZightMR(online~onlinegame+onlinevideo+onlinemusic,
                 data = census.at.school.5000, Labels=FALSE)
mroPara(mr2)
mr3 <- iNZightMR(online~onlinegame+onlinevideo+onlinemusic,
                 data = census.at.school.5000, Labels=letters[1:3])
mroPara(mr3)
}
}
\author{
Junjie Zeng
}

